backend/requirements.txt
fastapi
uvicorn[standard]
feedparser
psycopg2-binary
